+++ CHANGELOG.md
@@ -10,7 +10,7 @@ i.e. `<Major version>.<Minor version>.<Patch version>`
 ## [Unreleased] - TBA
 - Support call recovery capability (HA) #140
 - Support config build as var for kamailio/freeswitch
-
+- Gateway support specify outbound proxy
 
 ## [v0.7.1] - 2023-12-25
 - Fix UI access layer #121